requirements.txt
pyTelegramBotAPI
gspread
google-auth

